pyproject.toml
[project]
name = "code-reviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.13",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langgraph>=0.4.8",
    "pydantic>=2.11.7",
    "fastapi>=0.110.0",
    "uvicorn>=0.22.0",
    "PyJWT>=2.8.0",
    "requests>=2.31.0",
    "python-dotenv>=1.1.0",
]
